requirements.txt
kafka-python==2.0.0
PyJWT==1.7.1
